Move new stuff  to org.yakindu.sct.generator.core
