+++ plugins/org.yakindu.sct.generator.c/src/org/yakindu/sct/generator/c/CCodeGeneratorModule.java
@@ -25,10 +25,10 @@ import org.yakindu.sct.generator.c.eventdriven.EventDrivenStatemachineSource;
 import org.yakindu.sct.generator.c.language.CFunctionFactory;
 import org.yakindu.sct.generator.c.language.Function;
 import org.yakindu.sct.generator.c.types.CTypeSystemAccess;
-import org.yakindu.sct.generator.common.IFunction;
-import org.yakindu.sct.generator.common.factory.FunctionFactory;
 import org.yakindu.sct.generator.core.IExecutionFlowGenerator;
 import org.yakindu.sct.generator.core.IGeneratorModule;
+import org.yakindu.sct.generator.core.language.IFunction;
+import org.yakindu.sct.generator.core.language.factory.FunctionFactory;
 import org.yakindu.sct.generator.core.types.ICodegenTypeSystemAccess;
 import org.yakindu.sct.model.sexec.naming.INamingService;
 import org.yakindu.sct.model.sgen.FeatureParameterValue;
