--- plugins/org.yakindu.sct.generator.common/.gitignore
@@ -1,2 +1,0 @@
-/bin/
-/target/
