--- plugins/org.yakindu.sct.generator.common/META-INF/MANIFEST.MF
@@ -1,15 +1,0 @@
-Manifest-Version: 1.0
-Bundle-ManifestVersion: 2
-Bundle-Name: Yakindu SCT Generator Commons
-Bundle-SymbolicName: org.yakindu.sct.generator.common
-Bundle-Version: 3.1.0.qualifier
-Bundle-Vendor: statecharts.org
-Bundle-RequiredExecutionEnvironment: JavaSE-1.8
-Require-Bundle: com.google.guava,
- org.eclipse.xtext.xbase.lib,
- org.eclipse.xtend.lib,
- org.eclipse.xtend.lib.macro
-Export-Package: org.yakindu.sct.generator.common,
- org.yakindu.sct.generator.common.factory
-Import-Package: com.google.inject
-
