--- plugins/org.yakindu.sct.generator.common/build.properties
@@ -1,4 +1,0 @@
-source.. = src/
-output.. = bin/
-bin.includes = META-INF/,\
-               .
