--- plugins/org.yakindu.sct.generator.common/xtend-gen/.gitignore
@@ -1,5 +1,0 @@
-# Ignore those directories
-*
-
-# And never this file
-!.gitignore
