+++ plugins/org.yakindu.sct.generator.core/src/org/yakindu/sct/generator/core/language/ICodeBlock.java
@@ -8,20 +8,16 @@
  * 	rbeckmann - initial API and implementation
  * 
  */
-package org.yakindu.sct.generator.common
+package org.yakindu.sct.generator.core.language;
 
 /**
  * @author rbeckmann
  *
  */
-abstract class CodeBlock extends CodePart implements ICodeBlock {
-	protected CharSequence content;
+public interface ICodeBlock extends ICodePart {
+	public CharSequence getBlockOpen();
+	public CharSequence getBlockClose();
 	
-	override getContent() {
-		return content;
-	}
-
-	override setContent(CharSequence content) {
-		this.content = content;
-	}
+	public CharSequence getContent();
+	public void setContent(CharSequence content);
 }
