+++ plugins/org.yakindu.sct.generator.core/src/org/yakindu/sct/generator/core/language/ICodePart.java
@@ -8,15 +8,12 @@
  * 	rbeckmann - initial API and implementation
  * 
  */
-package org.yakindu.sct.generator.common;
-
-import java.util.List;
+package org.yakindu.sct.generator.core.language;
 
 /**
  * @author rbeckmann
  *
  */
-public interface IFunction extends ICodeBlock, IModifierOwner, ITypeOwner, INameOwner, IDocumentationOwner {
-	public List<CharSequence> getParameters();
-	public void setParameters(List<CharSequence> parameter);
+public interface ICodePart extends CharSequence {
+
 }
