+++ plugins/org.yakindu.sct.generator.core/src/org/yakindu/sct/generator/core/language/IParameter.java
@@ -8,12 +8,11 @@
  * 	rbeckmann - initial API and implementation
  * 
  */
-package org.yakindu.sct.generator.common;
+package org.yakindu.sct.generator.core.language;
 
 /**
  * @author rbeckmann
  *
  */
-public interface ITypeQualifier extends ICodePart {
-	
+public interface IParameter extends ITypeOwner, ITypeQualifierOwner, INameOwner {
 }
