+++ plugins/org.yakindu.sct.generator.core/src/org/yakindu/sct/generator/core/language/ITypeOwner.java
@@ -8,12 +8,13 @@
  * 	rbeckmann - initial API and implementation
  * 
  */
-package org.yakindu.sct.generator.common;
+package org.yakindu.sct.generator.core.language;
 
 /**
  * @author rbeckmann
  *
  */
-public interface IModifier extends ICodePart {
-
+public interface ITypeOwner extends ICodePart {
+	public IType getType();
+	public void setType(IType type);
 }
