+++ plugins/org.yakindu.sct.generator.core/src/org/yakindu/sct/generator/core/language/ITypeQualifier.java
@@ -8,13 +8,12 @@
  * 	rbeckmann - initial API and implementation
  * 
  */
-package org.yakindu.sct.generator.common;
+package org.yakindu.sct.generator.core.language;
 
 /**
  * @author rbeckmann
  *
  */
-public interface INameOwner extends ICodePart {
-	public CharSequence getName();
-	public void setName(CharSequence name);
+public interface ITypeQualifier extends ICodePart {
+	
 }
