+++ plugins/org.yakindu.sct.generator.core/src/org/yakindu/sct/generator/core/language/ITypeQualifierOwner.java
@@ -8,12 +8,13 @@
  * 	rbeckmann - initial API and implementation
  * 
  */
-package org.yakindu.sct.generator.common;
+package org.yakindu.sct.generator.core.language;
 
 /**
  * @author rbeckmann
  *
  */
-public interface ICodePart extends CharSequence {
-
+public interface ITypeQualifierOwner extends ICodePart {
+	public ITypeQualifier getTypeQualifier();
+	public void setTypeQualifier(ITypeQualifier typeQualifier);
 }
