+++ plugins/org.yakindu.sct.generator.core/src/org/yakindu/sct/generator/core/language/Parameter.xtend
@@ -8,7 +8,7 @@
  * 	rbeckmann - initial API and implementation
  * 
  */
-package org.yakindu.sct.generator.common
+package org.yakindu.sct.generator.core.language
 
 import org.eclipse.xtend.lib.annotations.Accessors
