+++ releng/org.yakindu.sct.releng/pom.xml
@@ -434,7 +434,6 @@
 		<module>../../plugins/org.yakindu.sct.ui.editor</module>
 		<module>../../plugins/org.yakindu.sct.ui.intro</module>
 		<module>../../plugins/org.yakindu.sct.ui.navigator</module>
-		<module>../../plugins/org.yakindu.sct.generator.common</module>
 		<module>../../plugins/org.yakindu.sct.generator.c</module>
 		<module>../../plugins/org.yakindu.sct.generator.cpp</module>
 		<module>../../plugins/org.yakindu.sct.generator.java</module>
